// src/types/actionUserTypes.tsx
export const SET_AUTO_COMPLETE_REGISTER = 'SET_AUTO_COMPLETE_REGISTER';
// ...
export interface SetAutocompleteRegisterAction {
  type: typeof SET_AUTO_COMPLETE_REGISTER;
  name: string;
  lastname: string;
}